third_party/blink/web_tests/external/wpt/css/css-flexbox/flex-one-sets-flex-basis-to-zero-px-expected.txt

This is a testharness.js-based test.
[FAIL] .flexbox 6
  assert_equals: \n<div class="flexbox column">\n  <div class="flex-half-one-zero-px" data-expected-height="4">Flex item with flex: 0.5 1 0px</div>\n</div>\nheight expected 4 but got 0
[FAIL] .flexbox 9
  assert_equals: \n<div class="flexbox column">\n  <div class="flex-one-one-zero-px" data-expected-height="14">Flex item with flex: 1 1 0px</div>\n</div>\nheight expected 14 but got 0
[FAIL] .flexbox 15
  assert_equals: \n<div class="flexbox column vertical">\n  <div class="flex-half-one-zero-px" data-expected-width="4">Flex item with flex: 0.5 1 0px</div>\n</div>\nwidth expected 4 but got 0
[FAIL] .flexbox 18
  assert_equals: \n<div class="flexbox column vertical">\n  <div class="flex-one-one-zero-px" data-expected-width="14">Flex item with flex: 1 1 0px</div>\n</div>\nwidth expected 14 but got 0
Harness: the test ran to completion.
